sumario/pyproject.toml
[project]
name = "sumario"
dynamic = ["version"]
description = "The best way to handle forms on any website."
authors = [
    {name = "JSTV Solutions", email = "sumario@jstvsolutions.com"},
]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "AGPL"}
dependencies = [
    "blinker==1.6.2",
    "email-validator==2.0.0.post2",
    "flask-babel==3.1.0",
    "flask-mail==0.9.1",
    "flask-migrate==4.0.4",
    "flask-sqlalchemy==3.0.5",
    "flask-user==1.0.2.2",
    "flask==2.3.3",
    "psycopg2-binary==2.9.7",
    "pyyaml==6.0.1",
    "sentry-sdk[flask]==1.30.0",
    "sqlalchemy-utils==0.41.1",
    "stripe==6.4.0",
]
requires-python = ">=3.11"

[project.urls]
Homepage = "https://sumar.io"

[project.optional-dependencies]
builder = [
    "bandit==1.7.5",
    "beautifulsoup4==4.12.2",
    "black==23.9.1",
    "build==1.0.3",
    "coverage==7.3.1",
    "html5lib==1.1",
    "pytest-cov==4.1.0",
    "pytest-env==1.0.1",
    "pytest==7.4.2",
    "safety==2.3.4",
    "twine==4.0.2",
    "wheel==0.41.2",
]

[tool.setuptools.dynamic]
version = {attr = "sumario.__version__"}
